+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- fixed GPIO implementation
+
 ### Removed
 
 ## [0.1.0] - 2024-08-16