+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- removed GPIO implementation
+
 ## [0.1.0] - 2024-08-16
 
 - initial version