+++ CHANGELOG.md
@@ -9,11 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `parl_io` Hub75 driver (untested)
+- `parl_io` Hub75 driver
 
 ### Changed
 
-- update `esp-hal` to 0.20.1
+- update `esp-hal` to 0.21.0
 - parl_io, lcd_cam: frequency is required to be passed in
 
 ### Fixed